Rekey the books table in the database
To allow the users to edit the ISBN the database needed to be rekeyed,
so when the ISBN field is edited it will no longer involve changing
the primary key of a record. To do add a new `id` column to be the new
key, and turn the ISBN column into a unique non-key column.

Also update the website to stop using ISBN values as the primary key
for books, but instead this new primary key.